+++ presenter-compose/impl/src/commonTest/kotlin/software/ralf/app/platform/presenter/compose/KotlinInjectInjectionTest.kt
@@ -5,29 +5,23 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import kotlin.coroutines.EmptyCoroutineContext
 import kotlin.test.Test
-import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.Inject
 import me.tatarka.inject.annotations.KmpComponentCreate
 import me.tatarka.inject.annotations.Provides
 import software.amazon.lastmile.kotlin.inject.anvil.AppScope
-import software.amazon.lastmile.kotlin.inject.anvil.ForScope
 import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
 import software.ralf.app.platform.presenter.PresenterCoroutineScope
-import software.ralf.app.platform.presenter.PresenterCoroutineScopeComponent
-import software.ralf.app.platform.scope.coroutine.MainCoroutineDispatcher
 
 class KotlinInjectInjectionTest {
 
   @Test
   fun `the PresenterCoroutineScope can be injected lazily`() {
     val testScope = CoroutineScope(CoroutineName("TestName"))
-    val testDispatcher = Dispatchers.Default
 
-    val component = createTestComponent(testScope, testDispatcher)
+    val component = createTestComponent(testScope)
 
     val composePresenterScope = component.composePresenterScopeFactory.createComposePresenterScope()
 
@@ -40,19 +34,12 @@ class KotlinInjectInjectionTest {
 
 @Component
 @SingleIn(AppScope::class)
-abstract class KotlinInjectTestComponent(
-  private val coroutineScope: CoroutineScope,
-  private val coroutineDispatcher: CoroutineDispatcher,
-) : PresenterCoroutineScopeComponent {
+abstract class KotlinInjectTestComponent(private val coroutineScope: CoroutineScope) {
   abstract val composePresenterScopeFactory: ComposePresenterScopeFactory
 
   @Provides
-  @ForScope(AppScope::class)
-  fun provideAppScopeCoroutineScope(): CoroutineScope = coroutineScope
-
-  @Provides
-  @MainCoroutineDispatcher
-  fun provideMainCoroutineDispatcher(): CoroutineDispatcher = coroutineDispatcher
+  @PresenterCoroutineScope
+  fun providePresenterCoroutineScope(): CoroutineScope = coroutineScope
 
   @Provides
   fun provideComposePresenterScopeFactory(
@@ -72,7 +59,4 @@ class KotlinInjectTestComposePresenterScopeFactory(
   )
 
 @KmpComponentCreate
-expect fun createTestComponent(
-  coroutineScope: CoroutineScope,
-  coroutineDispatcher: CoroutineDispatcher,
-): KotlinInjectTestComponent
+expect fun createTestComponent(coroutineScope: CoroutineScope): KotlinInjectTestComponent